src/main/java/com/whizat/study/springboot/web/HelloController.java
package com.whizat.study.springboot.web;

import com.whizat.study.springboot.dto.HelloResponseDto;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

@RestController // JSON을 반환하는 컨트롤로러 만들어줌. @ResponseBody를 각 메소드마다 선언했던것과 같음
public class HelloController {

    @GetMapping("/hello")
    public String hello() {
        return "hello";
    }

    @GetMapping("/hello/dto")   // @RequestMapping(mehtod = RequestMethod.GET)와 같음
    public HelloResponseDto helloDto(@RequestParam("name") String name, @RequestParam("amount") int amount) {
        return new HelloResponseDto(name, amount);
    }
}
